chore(build,rtc_*,ur5e_bringup): silence Tier 1 compiler warnings
build.sh: pass /opt/onnxruntime via CMAKE_PREFIX_PATH env var instead of
-D, so packages that don't find_package() onnxruntime no longer emit
"Manually-specified variables were not used" CMake warnings.

Source fixes (no behavior change):
- rtc_controller_manager: rmw_qos_profile_services_default →
  rclcpp::ServicesQoS() in CreateServices() (-Wdeprecated-declarations)
- rtc_tsid: mark unused n_vars in EomConstraint::compute_equality;
  static_cast<int>(cache.q.size()) in PostureTask (-Wunused-parameter,
  -Wconversion)
- rtc_base: check fscanf() return in thread_utils cgroup-v1 fallback
  (-Wunused-result); sentinel value preserved on read failure
- ur5e_bringup: drop {} brace-init on ControllerTopicHandles owned_topics_
  members and gui_msgs[] (avoids ROS msg explicit default ctor warning);
  check SetJointOrder() nodiscard result in InitHandModel + WARN on miss
- tests: (void)-cast nodiscard returns in warm-up / scenario steps
  (test_tsid_performance, test_exploration_motion)

All affected gtest binaries pass (39 binaries across 5 packages).

+++ rtc_tsid/src/tasks/posture_task.cpp
@@ -65,7 +65,7 @@ void PostureTask::compute_residual(const PinocchioCache& cache, const ControlRef
   // Note: q 공간에서의 차이는 nv 차원
   // Floating-base의 경우 q (nq) ≠ v (nv), pinocchio::difference 필요
   // 여기서는 간단히 head(nv) 사용 (fixed-base에서는 nq==nv)
-  const int nq = cache.q.size();
+  const int nq = static_cast<int>(cache.q.size());
   const int nv = nv_;
 
   if (nq == nv) {
